fix: disable app for mobiles
