+++ src/components/atoms/MobileMessage.js
@@ -1,0 +1,12 @@
+const MobileMessage = () => {
+  return (
+    <div>
+      <h2 className="text-error text-2xl text-center">
+        Please view this application on a larger screen.
+      </h2>
+      <p>Support for mobile screens coming soon! 🙃</p>
+    </div>
+  )
+}
+
+export default MobileMessage
